feat: group anyOf errors under each branch
It looks like there are a number of issues where people are complaining about how difficult it is to tell what the real errors are when a schema fails to match an `anyOf` constraint. I think the problem is that you can get two conflicting errors for two different branches, and there's nothing to indicate which conditions are "or" and which are "and".

My solution is a structure like:

```js
[ { keyword: 'anyOf',
    dataPath: '',
    schemaPath: '#/anyOf',
    params:
     { errors:
        [ [ { keyword: 'type',
              dataPath: '',
              schemaPath: '#/anyOf/0/type',
              params: { type: 'number' },
              message: 'should be number',
              schema: 'number',
              parentSchema: { type: 'number' },
              data: 'foo' } ],
          [ { keyword: 'type',
              dataPath: '',
              schemaPath: '#/anyOf/1/type',
              params: { type: 'integer' },
              message: 'should be integer',
              schema: 'integer',
              parentSchema: { type: 'integer' },
              data: 'foo' } ] ] },
    message: 'should match some schema in anyOf',
    schema: [ { type: 'number' }, { type: 'integer' } ],
    parentSchema: { anyOf: [ { type: 'number' }, { type: 'integer' } ] },
    data: 'foo' } ]
```

which indicates that one branch of the `anyOf` had the error "should be number" and the other branch had the error "should be integer".

This fails some of the tests in json-schema-test about recursive schemas, I'm not 100% sure how I should go about fixing those.

BREAKING CHANGE: the error structure has changed for `anyOf` errors.