Updated readme with proposed ReDoS solution
The solution being to run the format validation in a separate thread, with a timeout specified. If it doesn't complete within the timeout period, the data is assumed invalid, and the thread killed.

Granted, this breaks the synchronous nature of AJV validation, and is therefore outside of the library's own validation stack. But it might still be worth noting as a library-external solution to the problem.