make JSONSchemaType treat null as nullable
fixes #1664

JSONSchemaType now treats null as nullable instead of undefined,
allowing specifying non-null optional parameters.

They way that `Nullable<T>` handles types in `JSONSchemaType` makes it
very difficult to create `SomeJSONSchema` that any `JSONSchemaType` can
be assigned to. After many efforts to create this from scratch it
became clear that the only thing that would do this is a copy of
`JSONSchemaType` with `Nullable` fixed. Doing so revealed a lot of
surprising "bugs" in `SomeJSONSchema`, the most notable being some
places accept numbers where it should only accept string indices. This
seemed almost like a typescript bug, but I didn't delve in.

My conclusion from all of this is that it may be better to fully
redesign `JSONSchemaType` due to the number of inconsistencies. In light
of that, I opted for the least breaking change, which was actually to
hack `Nullable` to behave appropriately for `SomeJSONSchema`,
essentially, we check if that's what's being called, and if so, "do the
right thing". If someone were to try and type a copy of `Known` they
would similarly get a slightly more permissive type than they intended.
We could work around this by making the check explicit (akin to the
`IsPartial` keyword) but I think the times when this might happen aren't
bad enough to be worth it.