make ref and nullable top level
fixes #1652

A top level type couldn't be a ref, only items or properties. This
caused errors when an item or a property was SomeJSONSchema and
therefore would have $ref defined at a relative top level. It also
prevented types like `mySchema = { $ref: "..." }` which would be valid
if there were other registered schemas.

However this exposed another problem. In the past nullable was used to
mean undefineable, and so only made sense in properties (or kind of
items, but that's a little more suspect). Since nullable now means
contains a union with the type `null`, it also needed to be moved up a
level and not apply only to properties or items.

As a result, JSONPropertiesSchema didn't really serve a purpose and so
was removed.